renamed 'routes' to 'pages'
